re-factored job-scheduler file
